% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfillCrits.R
\name{as.MBOInfillCrit}
\alias{as.MBOInfillCrit}
\title{Conver to MBOInfillCrit

Necessary to use \code{mlrMBO::setMBOControlInfill()}.}
\usage{
as.MBOInfillCrit(x)
}
\arguments{
\item{x}{\code{\link{Infill}} the Infill object to convert}
}
\value{
\code{MBOInfillCrit}
}
\description{
Conver to MBOInfillCrit

Necessary to use \code{mlrMBO::setMBOControlInfill()}.
}
\seealso{
Other Utility Functions: \code{\link{getNugget}}

Other Infill Criteria: \code{\link{InfillCB}},
  \code{\link{acquisition}}
}
\concept{Infill Criteria}
\concept{Utility Functions}
