% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectResults.R
\name{collectMBFResult}
\alias{collectMBFResult}
\title{Collect Result Information}
\usage{
collectMBFResult(opt.state, aggregate.perresult = list(domHV =
  function(x) computeHV(x, ref.point)),
  ref.point = opt.state$control$nadir)
}
\arguments{
\item{opt.state}{\code{[MBOSingleObjResult]} \code{mobafeasMBO()} result to analyse}

\item{aggregate.perresult}{\code{[list]} list of functions
to apply to fitness and holdout fitness. Every entry must either be
a \code{character(1)} naming the function to use, or a function, in which
that entry must have a name. Each function must return exactly one
numeric value when fed a fitness matrix of one generation.}

\item{ref.point}{\code{[numeric(2)]} reference point to use for HV computation}
}
\value{
\code{data.frame}
}
\description{
Collect the \code{MBOSingleObjResult} information for a complete \code{data.frame}
with information about progress (both on training data and on
holdout data) and ressource usage.
}
