% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMobafeasLearner.R
\name{constructMBFLearner}
\alias{constructMBFLearner}
\title{Construct MoBaFeaS Gaussian Process Learner}
\usage{
constructMBFLearner(par.set, kernelconstructor = NULL,
  default.kernel = "matern3_2")
}
\arguments{
\item{par.set}{\code{[ParamSet]} The parameter set of the objective function being optimized
(probably the return value of \code{makeMobafeasObjective()}).}

\item{kernelconstructor}{\code{[function | NULL]} function \code{d -> kernel}, see the \code{\link{Kernel}} functions.}

\item{default.kernel}{\code{[character(1)]} One of \code{"matern3_2"} (default), \code{"matern5_2"}, \code{"exp"}, \code{"gauss"}.}
}
\value{
\code{Learner}.
}
\description{
The mobafeas Learner is specifically set up to handle feature configuration vectors
with a separate kernel.
}
\seealso{
Other Surrogate Models: \code{\link{constructRFSurrogate}}
}
\concept{Surrogate Models}
