% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMobafeasLearner.R
\name{constructRFSurrogate}
\alias{constructRFSurrogate}
\title{Construct the default Random Forest surrogate learner for mlrMBO}
\usage{
constructRFSurrogate()
}
\description{
Constructs the \code{randomForest} \code{Learner} that would be used by \code{mlrMBO} if the decision space
has discrete components.
}
\seealso{
Other Surrogate Models: \code{\link{constructMBFLearner}}
}
\concept{Surrogate Models}
