% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infillOpt.R
\name{getNugget}
\alias{getNugget}
\title{Get Nugget from Model}
\usage{
getNugget(learner.model)
}
\arguments{
\item{learner.model}{\code{[model object]} the model to query.}
}
\value{
\code{numeric(1)} nugget value.
}
\description{
Return the nugget value of the (Gaussian process) model.

Currently defined for \code{DiceKriging} and \code{kergp} models.
}
\seealso{
Other Utility Functions: \code{\link{as.MBOInfillCrit}}
}
\concept{Utility Functions}
