% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfillCrits.R
\name{Infill}
\alias{Infill}
\alias{InfillCB}
\alias{InfillEI}
\alias{InfillResponse}
\title{Infill Criterion

Different Infill criteria. Supported are:
\itemize{
\item Confidence Bound (\code{InfillCB})
\item Expected Improvement (\code{InfillEI})
\item Response (\code{InfillResponse})
}}
\usage{
InfillCB(lambda = 2)

InfillEI()

InfillResponse()
}
\arguments{
\item{lambda}{\code{[numeric(1)]}}
}
\description{
Infill Criterion

Different Infill criteria. Supported are:
\itemize{
\item Confidence Bound (\code{InfillCB})
\item Expected Improvement (\code{InfillEI})
\item Response (\code{InfillResponse})
}
}
\seealso{
Other Infill Criteria: \code{\link{acquisition}},
  \code{\link{as.MBOInfillCrit}}
}
\concept{Infill Criteria}
