% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infillOpt.R
\name{infillOptMobafeas}
\alias{infillOptMobafeas}
\title{MOSMAFS-based Multi-Crit Infill-Optimization}
\usage{
infillOptMobafeas(infill.crit, models, control, par.set, opt.path, designs,
  iter, ...)
}
\arguments{
\item{infill.crit}{\code{[Infill]} Infill criterion being used.}

\item{models}{\code{[list of WrappedModel]} Model(s) trained on the performance data so far. Only the
first model is considered.}

\item{control}{\code{[MBOControl]} \code{mlrMBO} control object, as augmented by \code{makeMBFControl()} and
\code{mobafeasMBO()}.}

\item{par.set}{\code{[ParamSet]} decision space over which MBO-optimization is performed.}

\item{opt.path}{\code{[OptPath]} OptPath of results so far}

\item{designs}{\code{[list of data.frame]} data.frame of results so far}

\item{iter}{\code{[integer(1)]} iteration number. Ignored.}
}
\value{
\code{data.frame} of suggested point to evaluate.
}
\description{
Uses \code{\link[mosmafs:mosmafs-package]{mosmafs::mosmafs-package}} to perform infill optimization. The infill criteria are not
the ones from \code{mlrMBO}, but instead the ones given by an \code{\link{Infill}} object.

This function should probably not be called directly, but instead should be given to \code{mlrMBO::mbo}
as the infill criterion. This is usually done using \code{makeMBFControl()} and then \code{mobafeasMBO()}.
}
