% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{Kernel}
\alias{Kernel}
\alias{kernelMBFHamming}
\alias{kernelMBFGraph}
\alias{kernelMBFAgreement}
\alias{kernelMBFAgreeCor}
\title{Feature Selection Kernels}
\usage{
kernelMBFHamming()

kernelMBFGraph(allequal = TRUE)

kernelMBFAgreement(limit.par = FALSE)

kernelMBFAgreeCor(data, limit.par = FALSE)
}
\arguments{
\item{allequal}{\code{[logical(1)]} Whether to generate one parameter per feature, or only one parameter
overall.}

\item{limit.par}{\code{[logical(1)]} Whether to limit the parameter value range}

\item{data}{\code{[matrix | data.frame | Task]} the data to use.}
}
\description{
Kernels defined on the space of possible feature selections. These should be run, possibly with
arguments, and the result given to \code{constructMBFLearner()}.
}
