% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMBFControl.R
\name{makeMBFControl}
\alias{makeMBFControl}
\title{MoBaFeaS Configuration Object}
\usage{
makeMBFControl(mosmafs.config, propose.points = 1,
  infill.crit = InfillEI())
}
\arguments{
\item{mosmafs.config}{\code{[MosmafsConfig]} An object created by \code{MosmafsConfig()}.}

\item{propose.points}{\code{[integer(1)]} How many points to propose per model fit. Currently
only supports 1.}

\item{infill.crit}{\code{[Infill]} An \code{\link{Infill}} object to use.}
}
\value{
\code{MBOControl}
}
\description{
Configure the \code{mobafeasMBO()} run. This uses a \code{mlrMBO::makeMBOControl()} object and
sets it up for the specific needs of mobafeas. The resulting object should still have
an adequate terminator set using \code{mlrMBO::setMBOControlTermination()}.
}
\seealso{
Other Control Objects: \code{\link{MosmafsConfig}},
  \code{\link{makeMobafeasObjective}},
  \code{\link{setMBFControlInfill}}
}
\concept{Control Objects}
