% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo.R
\name{mobafeasMBO}
\alias{mobafeasMBO}
\title{Start mobafeas MBO

A wrapper around \code{mlrMBO::mbo()} that performs the mobafeas-specific black-box / white-box two-criteria
optimization for feature selection.}
\usage{
mobafeasMBO(fun, population,
  learner = constructMBFLearner(getParamSet(fun)), control,
  show.info = getOption("mlrMBO.show.info", TRUE), more.args = list())
}
\arguments{
\item{fun}{\code{[smoof_function]} smoof function with additional attributes \code{"co.objective"} and \code{"nadir"}.
\code{"co.objective"} is a \code{function} taking the design \code{data.frame} and returning a vector of second-objective-values.
\code{"nadir"} is a \code{numeric(2)} nadir-point for objective 1 (of the \code{fun} return value) and 2 (the \code{"co.objective"}
return value). This should probably be generated using \code{makeMobafeasObjective()}.}

\item{population}{\code{[list]} list of parameter values for initial design.}

\item{learner}{\code{[Learner | NULL]} an \code{mlr} \code{Learner} object, usually created using \code{constructMBFLearner()} or
\code{constructRFSurrogate()}. Giving \code{NULL} uses the \code{mlrMBO} default learner (a \code{"regr.km"} or \code{"regr.randomForest"},
depending on whether the decision space has discrete variables, configured for mbo).}

\item{control}{\code{[MBOControl]} A control object, generated by \code{makeMBFControl()}}

\item{show.info}{\code{[logical(1)]} whether to output run progress information.}

\item{more.args}{\code{[list]} arguments to give to the objective function. Should be an empty list in most cases.}
}
\value{
\code{MBOSingleObjResult} the optimization result, should probably be analyzed using \code{collectMBFResult()}.
}
\description{
Start mobafeas MBO

A wrapper around \code{mlrMBO::mbo()} that performs the mobafeas-specific black-box / white-box two-criteria
optimization for feature selection.
}
