% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMBFControl.R
\name{setMBFControlInfill}
\alias{setMBFControlInfill}
\title{Change the Configuration Object Infill Optimization}
\usage{
setMBFControlInfill(control, infill.crit = NULL, mosmafs.config = NULL)
}
\arguments{
\item{control}{\code{[MBOControl]} Control object created with \code{makeMBFControl()}.}

\item{infill.crit}{\code{[Infill]} An \code{\link{Infill}} object to use.}

\item{mosmafs.config}{\code{[MosmafsConfig]} An object created by \code{MosmafsConfig()}.}
}
\value{
\code{MBOControl}
}
\description{
Set the infill optimization config or infill criterion of the \code{makeMBFControl()}
generated control object.
}
\seealso{
Other Control Objects: \code{\link{MosmafsConfig}},
  \code{\link{makeMBFControl}},
  \code{\link{makeMobafeasObjective}}
}
\concept{Control Objects}
